remove legacy version
