--- Mohr_Circle_legacy/description.html
@@ -1,57 +1,0 @@
-<style>
-h1 {
-    margin: 2em 0 0 0;
-    color: #2e484c;
-    font-family: 'Julius Sans One', sans-serif;
-    font-size: 1.8em;
-    text-transform: uppercase;
-}
-a:link {
-    font-weight: bold;
-    text-decoration: none;
-    color: #0d8ba1;
-}
-a:visited {
-    font-weight: bold;
-    text-decoration: none;
-    color: #1a5952;
-}
-a:hover, a:focus, a:active {
-    text-decoration: underline;
-    color: #9685BA;
-}
-p {
-    font: "Libre Baskerville", sans-serif;
-    text-align: justify;
-    text-justify: inter-word;
-}
-code {
-    color: #336699;
-}
-</style>
-
-<h1>Mohr's Circle</h1>
-
-<p>
-This app visualizes Mohr's Circle.
-It explains how Mohr's Circle can be used to identify the different stresses that act on a differential element with variable orientation.
-</p>
-
-<p>
-The principal stresses can be computed with
-$$\sigma_{\mathrm{I}/\mathrm{II}} = \frac{1}{2} \, \left[ \sigma_z + \sigma_x \pm \sqrt{\left( \sigma_z - \sigma_x \right)^2 + 4 \, \tau_{xz}^2} \right],$$
-the principal stress angle with
-$$ \tan \left( 2 \, \alpha_0 \right) = 2 \cdot \frac{\tau_{xz}}{\sigma_z - \sigma_x}.$$
-</p>
-
-<p>
-In this application, the normal stresses in $x$- and in $z$-direction $\sigma_x$ and $\sigma_z$ can be changed, as well as the tangential stress $\tau_{xz}$.
-</p>
-
-<p>
-After selecting the initial stress state, a click on "Draw" shows Mohr's Circle. Visualized are the normal stresses $\sigma_x$ and $\sigma_z$, the shear stress $\tau_{xz}$, and the current stress state $A$. A click on "Show principal stress + direction" reveals the principal stresses $\sigma_I$ and $\sigma_{II}$, the center of Mohr's Circle $\sigma_M$, and the principal stress angle $\alpha_0$.</p>
-
-<p>
-The element can be rotated in the third figure, resulting in the stress state $B$. It can be observed, that the shear stresses vanish, if the element is rotated according to $\alpha_0$. The current normal stresses $\sigma_{\overline{x}}$ and $\sigma_{\overline{z}}$ can also be seen in the plot visualizing Mohr's Circle.
-
-</p>
